[chore] Refactor Unit Test package
